requirements.txt
beautifulsoup4
google-cloud-firestore
google-generativeai
pandas
python-dotenv
requests
selenium
streamlit
webdriver-manager
